app/tools/search_tools.py
import asyncio
import httpx
from typing import List, Dict, Any, Optional
from bs4 import BeautifulSoup
import json

from app.models.schemas import SearchResult
from app.core.config import settings

class WebSearchTool:
    """Tool for performing web searches"""
    
    def __init__(self):
        self.client = httpx.AsyncClient(timeout=settings.SEARCH_TIMEOUT)
        
    async def search(self, query: str) -> List[SearchResult]:
        """
        Perform a web search and return results
        
        Note: This is a mock implementation. In production, you would:
        1. Use a real search API (Google, Bing, etc.)
        2. Implement proper rate limiting
        3. Handle API errors gracefully
        """
        
        # Mock search results for demonstration
        # In production, replace with actual API calls
        mock_results = await self._mock_search(query)
        
        # Process results in parallel
        tasks = [self._fetch_content(result) for result in mock_results]
        enriched_results = await asyncio.gather(*tasks, return_exceptions=True)
        
        # Filter out failed fetches
        valid_results = [
            r for r in enriched_results 
            if isinstance(r, SearchResult)
        ]
        
        return valid_results
        
    async def _mock_search(self, query: str) -> List[Dict[str, Any]]:
        """Mock search function for demonstration"""
        
        # In production, this would call a real search API
        # For now, return mock data based on query keywords
        
        # Create more realistic mock results based on query
        if "ai" in query.lower() or "agent" in query.lower():
            base_results = [
                {
                    "url": "https://anthropic.com/research/multi-agent-systems",
                    "title": f"Multi-Agent AI Systems: {query} Research",
                    "snippet": f"Comprehensive analysis of {query} showing how multi-agent systems outperform single agents by 90% in research tasks..."
                },
                {
                    "url": "https://openai.com/blog/ai-agents-2025",
                    "title": f"OpenAI's Latest AI Agents for {query}",
                    "snippet": f"OpenAI releases new AI agents specifically designed for {query} with improved reasoning capabilities..."
                },
                {
                    "url": "https://arxiv.org/abs/2024.12345",
                    "title": f"Academic Paper: {query} in Modern AI",
                    "snippet": f"Peer-reviewed research on {query} demonstrates significant improvements in accuracy and efficiency..."
                },
                {
                    "url": "https://techcrunch.com/2025/ai-breakthrough",
                    "title": f"TechCrunch: AI Breakthrough in {query}",
                    "snippet": f"Industry analysis reveals that {query} represents a major advancement in artificial intelligence..."
                },
                {
                    "url": "https://nature.com/articles/ai-research-2025",
                    "title": f"Nature: Scientific Study on {query}",
                    "snippet": f"Scientific publication examining the implications of {query} for future AI development..."
                }
            ]
        else:
            # Generic results for other queries
            base_results = [
                {
                    "url": f"https://example.com/research-{i}",
                    "title": f"Research Article: {query} - Study {i}",
                    "snippet": f"This comprehensive study examines {query} and provides detailed analysis of current trends and findings..."
                }
                for i in range(1, 6)
            ]
        
        return base_results
        
    async def _fetch_content(self, result: Dict[str, Any]) -> SearchResult:
        """Fetch and parse content from a URL"""
        
        try:
            # In production, actually fetch the URL
            # For mock, just create a SearchResult with expanded content
            
            # Generate more realistic content based on the title and snippet
            expanded_content = f"""
            {result['snippet']}
            
            This article provides in-depth coverage of the topic, including:
            - Historical context and background information
            - Current state of research and development
            - Key findings from recent studies
            - Expert opinions and industry analysis
            - Future implications and recommendations
            
            The research methodology involved comprehensive data collection and analysis,
            ensuring the reliability and validity of the presented findings.
            """
            
            return SearchResult(
                url=result["url"],
                title=result["title"],
                snippet=result["snippet"],
                content=expanded_content,
                relevance_score=0.8  # Will be updated by agent
            )
            
        except Exception as e:
            print(f"Error fetching {result.get('url', 'unknown')}: {e}")
            raise
            
    async def close(self):
        """Clean up resources"""
        await self.client.aclose()