requirements.txt
fastapi
uvicorn
pydantic
httpx
beautifulsoup4
litellm
anthropic
python-dotenv
redis
asyncio